+++ packages/kilo-jetbrains/backend/src/main/kotlin/ai/kilocode/server/KiloProcessState.kt
@@ -1,0 +1,6 @@
+package ai.kilocode.server
+
+sealed class KiloProcessState {
+    data class Ready(val port: Int, val password: String) : KiloProcessState()
+    data class Error(val message: String, val details: String? = null) : KiloProcessState()
+}
