+++ packages/kilo-jetbrains/backend/src/main/resources/kilo.jetbrains.backend.xml
@@ -6,6 +6,6 @@
 
     <extensions defaultExtensionNs="com.intellij">
         <applicationService
-            serviceImplementation="ai.kilocode.jetbrains.KiloManager"/>
+            serviceImplementation="ai.kilocode.server.KiloCliService"/>
     </extensions>
 </idea-plugin>
